refactor: 구조 및 타입 수정
