--- app/(auth)/login/_entities/model/types.ts
@@ -1,1 +1,0 @@
-export type LoginProvider = "GOOGLE" | "KAKAO";
